ml-service/pyproject.toml
[project]
name = "stableguard-ml-service"
version = "0.1.0"
description = "Standalone ML inference service for StableGuard"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "pydantic==2.11.7",
    "pydantic-settings==2.7.0",
    "python-multipart==0.0.20",
    "Pillow==11.3.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.45.0",
    "accelerate>=0.30.0",
    "huggingface-hub>=0.25.0",
    "qwen-vl-utils>=0.0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
